now do the revert, but the other change also makes sense

+++ src/gt4py/next/program_processors/codegens/gtfn/codegen.py
@@ -59,7 +59,7 @@ class GTFNCodegen(codegen.TemplatedGenerator):
         "float": "double",
         "float32": "float",
         "float64": "double",
-        "int": "int",
+        "int": "long",
         "int32": "std::int32_t",
         "int64": "std::int64_t",
         "bool": "bool",
